requirements.txt
flask==2.3.3
flask-cors==4.0.0
Flask==2.3.3
telethon==1.29.3
google-api-python-client==2.105.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
